requirements.txt
decorator==4.0.2
ipython==4.0.0
ipython-genutils==0.1.0
matplotlib==1.4.3
nltk==3.0.5
numpy==1.9.2
Optunity==1.1.1
pandas==0.16.2
path.py==8.1
pickleshare==0.5
pyenchant==1.6.6
pyparsing==2.0.3
python-dateutil==2.4.2
pytz==2015.4
requests==2.7.0
scikit-learn==0.16.1
scipy==0.16.0
seaborn==0.6.0
simplegeneric==0.8.1
six==1.9.0
traitlets==4.0.0
twitter==1.17.1
Whoosh==2.7.0